# Conflicts:
#	sampledata/src/main/java/com/jjg/game/sampledata/GameDataManager.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/AuxiliaryAwardCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/AvatarCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BaseElementCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BaseElementRewardCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BaseInitCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BaseLineCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BaseLineFreeCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BaseRollerCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BaseRollerModeCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BaseRoomCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BetAreaCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BetRobotCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/BlackjackCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/ChessRobotCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/ClientFreeRollerCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/ClientRollerCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/ConditionCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/GameFunctionCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/GameListCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/GlobalConfigCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/ItemCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/MailCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/PlayerLevelConfigCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/PokerPoolCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/PoolCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/RobotActionCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/RobotCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/RoomCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/RoomExpendCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/Room_BetCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/Room_ChessCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/ShopConfigCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/SouthernMoneyCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/SpecialAuxiliaryAwardCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/SpecialAuxiliaryCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/SpecialGirdCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/SpecialPlayCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/SpecialResultLibCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/StatusCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/TexasCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/WarehouseCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/bean/WinPosWeightCfg.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/AuxiliaryAwardCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/AvatarCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BaseElementCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BaseElementRewardCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BaseInitCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BaseLineCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BaseLineFreeCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BaseRollerCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BaseRollerModeCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BaseRoomCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BetAreaCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BetRobotCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/BlackjackCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/ChessRobotCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/ClientFreeRollerCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/ClientRollerCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/ConditionCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/GameFunctionCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/GameListCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/GlobalConfigCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/ItemCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/MailCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/PlayerLevelConfigCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/PokerPoolCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/PoolCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/RobotActionCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/RobotCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/RoomCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/RoomExpendCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/Room_BetCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/Room_ChessCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/ShopConfigCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/SouthernMoneyCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/SpecialAuxiliaryAwardCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/SpecialAuxiliaryCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/SpecialGirdCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/SpecialPlayCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/SpecialResultLibCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/StatusCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/TexasCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/WarehouseCfgContainer.java
#	sampledata/src/main/java/com/jjg/game/sampledata/container/WinPosWeightCfgContainer.java

+++ core/src/main/java/com/jjg/game/core/data/Item.java
@@ -9,12 +9,12 @@ public class Item {
     //道具id
     private int id;
     //道具数量
-    private int count;
+    private long count;
 
     public Item() {
     }
 
-    public Item(int id, int count) {
+    public Item(int id, long count) {
         this.id = id;
         this.count = count;
     }
@@ -27,15 +27,15 @@ public class Item {
         this.id = id;
     }
 
-    public int getCount() {
+    public long getCount() {
         return count;
     }
 
-    public void setCount(int count) {
+    public void setCount(long count) {
         this.count = count;
     }
 
-    public void addCount(int count) {
+    public void addCount(long count) {
         this.count += count;
     }
 }
